# Runtime compressed refs work