--- a/runtime/compiler/ilgen/Walker.cpp
+++ b/runtime/compiler/ilgen/Walker.cpp
@@ -2150,8 +2150,6 @@ TR_J9ByteCodeIlGenerator::calculateArrayElementAddress(TR::DataType dataType, bo
    // since each element of an reference array is a compressed pointer,
    // modify the width accordingly, so the stride is 4bytes instead of 8
    //
-   // OMR_GC_COMPRESSED_POINTERS
-   //
    if (comp()->useCompressedPointers() && dataType == TR::Address)
       {
       width = TR::Compiler->om.sizeofReferenceField();