… runtime/compiler/x/codegen/J9TreeEvaluator.cpp
… runtime/compiler/x/runtime/X86LockReservation.nasm
… runtime/compiler/x/runtime/X86Unresolveds.nasm
… runtime/compiler/z/codegen/J9TreeEvaluator.cpp
… runtime/jilgen/jilconsts.c
… runtime/oti/j9accessbarrier.h