Get the X JIT going in mixed mode.

[ci skip]

Signed-off-by: Graham Chapman <graham_chapman@ca.ibm.com>