@andrewcraik 2019-09-03
I am still reading through the changes - @gacholio what performance evaluation has been done? any assessment of compile-time impact from making the compression query non-constant?