@gacholio 2019-09-03
The assumption is that `if (true)` or `if (false)` will be removed by the compiler, so there should be zero perf impact on the normal builds. I have not verified it.

The assembly code is identical in the normal builds as well - the check is only done in mixed mode.