@gacholio 2019-09-03
No one has noted a perf regression with earlier changes to VM/GC/JIT (though again, it has not been expressly measured).