@andrewcraik 2019-09-04
FYI @fjeremic, @zl-wang and @0xdaryl since there are codegen changes for z, p and arm codegens that also need to be reviewed.