@andrewcraik 2019-09-06
I've finished going through and @fjeremic caught the things I was thinking through. Compile-time perf evaluation will be needed before I can approve the change since there is a risk of regressions with the extra path length required in the compiler for the mixed build.