@gacholio 2019-09-06
Since this has passed one review, I'll prepare some builds for perf testing next week. Any requested changes are unlikely to make or break performance.