@gacholio 2019-09-09
Looks like my failing test build was using omr master, not the openj9 accepted version, so I picked up an unpromoted jprofiling change, so I believe this PR actually has no effect on the problem. The fact that the jprofiling calls are not using the right linkage is a separate issue.