@DanHeidinga 2019-09-24
@gacholio @andrewcraik Are the perf results for this available?  Is there anything else that needs to be done before this can be merged?