@andrewcraik 2019-09-26
IBM was kind enough to do some perf runs on their performance setups to help verify the compile time and throughput of this change. Summarizing on behalf of @piyush286 who organized the runs. These runs were done on a modern linux x86 box using the DayTrader 7 benchmark, 6 iterations of each configuration with runs interwoven:

metric | baseline | runtime compressedrefs checks | Notes 
-------|---------|-----------------------------------|-------
startup | 4515.4ms CI 0.38% | 4486.8ms CI 0.35% | No regression 
footprint | 228868.8kb CI 1.24% | 228370.3kb CI 1.06% | No regression 
JIT CPU time cold | 145287.3ms CI 1.56%  | 146911.3ms CI 1.68% | No regression 
JIT CPU time warm | 114358.7ms CI 1.56% | 114109.2ms CI 2.87% | No regression 

Since DayTrader is a large app framework where no method dominates in the profile and many compiles need to be done quickly the lack of regression here says that there should be no performance impact from these changes.

Looking at the code I think that the changes are correct - there are a few bits that needed some thought in the assembly helpers, but I think it makes sense.