@DanHeidinga 2019-10-10
@andrewcraik builds have passed.  Do you want to merge this?